改成單一房間指定receiver寫法
